+++ src/class/flagship/flagship.ts
@@ -2,7 +2,15 @@ import { FsLogger } from '@flagship.io/js-sdk-logs';
 import { EventEmitter } from 'events';
 
 import defaultConfig, { internalConfig } from '../../config/default';
-import { FlagshipSdkConfig, IFlagship, IFlagshipBucketing, IFlagshipVisitor, IFsPanicMode, IFsCacheManager } from '../../types';
+import {
+    FlagshipSdkConfig,
+    IFlagship,
+    IFlagshipBucketing,
+    IFlagshipVisitor,
+    IFsPanicMode,
+    IFsCacheManager,
+    NewVisitorOptions
+} from '../../types';
 import flagshipSdkHelper from '../../lib/flagshipSdkHelper';
 import loggerHelper from '../../lib/loggerHelper';
 import Bucketing from '../bucketing/bucketing';
@@ -36,7 +44,10 @@ class Flagship implements IFlagship {
         this.bucket = null;
         this.panic = new PanicMode(this.config);
         this.envId = envId;
-        this.cacheManager = utilsHelper.isClient() ? clientCacheManager : null;
+        this.cacheManager = null;
+        if (this.config.enableClientCache && utilsHelper.isClient()) {
+            this.cacheManager = clientCacheManager;
+        }
         if (!apiKey) {
             this.log.warn(
                 'WARNING: "start" function signature will change in the next major release. "start(envId, settings)" will be "start(envId, apiKey, settings)", please make this change ASAP!'
@@ -67,7 +78,11 @@ class Flagship implements IFlagship {
         }
     }
 
-    public newVisitor(id: string, context: FlagshipVisitorContext): IFlagshipVisitor {
+    public newVisitor(id: string, context: FlagshipVisitorContext, options: NewVisitorOptions = {}): IFlagshipVisitor {
+        const defaultOptions: NewVisitorOptions = {
+            isAuthenticated: undefined
+        };
+        const { isAuthenticated } = { ...defaultOptions, ...options };
         const logBook = {
             API: {
                 newVisitorInfo: `new visitor (id="${id}") calling decision API for initialization (waiting to be ready...)`,
@@ -84,6 +99,7 @@ class Flagship implements IFlagship {
         const flagshipVisitorInstance = new FlagshipVisitor(this.envId, id, this.panic, this.config, {
             bucket: this.bucket,
             context,
+            isAuthenticated,
             cacheManager: this.cacheManager
         });
         this.log.info(`Creating new visitor (id="${flagshipVisitorInstance.id}")`);
@@ -141,12 +157,21 @@ class Flagship implements IFlagship {
      * [When fetchNow/activateNow=true] Update visitor will check if a synchronize is needed based on visitor context changed or no modifications have been fetched before, then it will emit "ready" event.
      * [When both fetchNow/activateNow=false] It will just emit "ready" event, without any change.
      */
-    public updateVisitor(visitorInstance: IFlagshipVisitor, context: FlagshipVisitorContext): IFlagshipVisitor {
+    public updateVisitor(
+        visitorInstance: IFlagshipVisitor,
+        payload: { context?: FlagshipVisitorContext; isAuthenticated?: boolean }
+    ): IFlagshipVisitor {
+        const defaultPayload = {
+            context: visitorInstance.context,
+            isAuthenticated: visitorInstance.isAuthenticated
+        };
+        const { context, isAuthenticated } = { ...defaultPayload, ...payload };
         this.log.debug(`updateVisitor - updating visitor (id="${visitorInstance.id}")`);
         const flagshipVisitorInstance = new FlagshipVisitor(this.envId, visitorInstance.id, this.panic, this.config, {
             bucket: this.bucket,
             previousVisitorInstance: visitorInstance,
             context,
+            isAuthenticated,
             cacheManager: this.cacheManager
         });
 
@@ -195,11 +220,11 @@ class Flagship implements IFlagship {
         }
         if (this.bucket !== null && this.bucket.isPollingRunning) {
             this.log.warn(
-                `startBucketingPolling - bucket already polling with interval set to "${this.config.pollingInterval}" minute(s).`
+                `startBucketingPolling - bucket already polling with interval set to "${this.config.pollingInterval}" second(s).`
             );
             return {
                 success: false,
-                reason: `startBucketingPolling - bucket already polling with interval set to "${this.config.pollingInterval}" minute(s).`
+                reason: `startBucketingPolling - bucket already polling with interval set to "${this.config.pollingInterval}" second(s).`
             };
         }
         this.log.error('startBucketingPolling - bucket not initialized, make sure "decisionMode" is set to "Bucketing"');